Add Publisher field to the Book domain class
A new Publisher field has been introduced to the Book class to link a publisher to a book. Getter and setter methods for this field have also been added. Additionally, the Publisher and Book classes have been updated to reflect this relationship, and the BootStrapData class has been modified to use this new Publisher field.